# openapi.yaml
openapi: "3.1.0"
info:
  title: Malluland API
  version: "1.0.0"
  description: Official API for Malluland backend services.
servers:
  - url: http://localhost:8787
    description: Local server
components:
  securitySchemes:
    XUserId:
      type: apiKey
      in: header
      name: x-user-id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
  responses:
    UnauthorizedHeader:
      description: Missing x-user-id header
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
paths:
...
  /meetups:
    get:
      summary: List/discover meetups
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: filter
          in: query
          schema:
            type: string
            enum: [upcoming, this-week, this-weekend]
        - name: city
          in: query
          schema:
            type: string
        - name: activityId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Meetups list
          content:
            application/json:
              schema:
                type: object
                properties:
                  page: { type: integer }
                  pageSize: { type: integer }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        startsAt: { type: string, format: date-time }
                        endsAt: { type: string, format: date-time, nullable: true }
                        city: { type: string, nullable: true }
                        state: { type: string, nullable: true }
                        country: { type: string, nullable: true }
                        activityId: { type: integer }
                        whoPays: { type: string, nullable: true }
                        feeAmount: { type: string, nullable: true }
                        guests: { type: integer }
                        hostId: { type: string, format: uuid }
                        hostName: { type: string, nullable: true }
                        hostAvatar: { type: string, nullable: true }
    post:
      summary: Create a meetup
      security:
        - XUserId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, activityId, guests, whoPays, currencyCode, feeAmount, locationText, description, startsAt, endsAt, city, state, country]
              properties:
                name: { type: string }
                activityId: { type: integer }
                guests: { type: integer }
                whoPays: { type: string }
                currencyCode: { type: string }
                feeAmount: { type: string }
                locationText: { type: string }
                description: { type: string }
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time }
                mapUrl: { type: string, nullable: true }
                city: { type: string }
                state: { type: string }
                country: { type: string }
                lat: { type: number, nullable: true }
                lng: { type: number, nullable: true }
      responses:
        '201':
          description: Created meetup
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/UnauthorizedHeader'
  /meetups/me:
    get:
      summary: List meetups hosted by the acting user
      security:
        - XUserId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: includePast
          in: query
          schema: { type: string, enum: [true, false] }
      responses:
        '200':
          description: My meetups list
          content:
            application/json:
              schema:
                type: object
                properties:
                  page: { type: integer }
                  pageSize: { type: integer }
                  items: { type: array, items: { type: object } }
  /meetups/{id}:
    get:
      summary: Get meetup by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Meetup details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  startsAt: { type: string, format: date-time }
                  endsAt: { type: string, format: date-time, nullable: true }
                  city: { type: string, nullable: true }
                  state: { type: string, nullable: true }
                  country: { type: string, nullable: true }
                  activityId: { type: integer }
                  whoPays: { type: string, nullable: true }
                  feeAmount: { type: string, nullable: true }
                  guests: { type: integer }
                  hostId: { type: string, format: uuid }
                  hostName: { type: string, nullable: true }
                  hostAvatar: { type: string, nullable: true }
                  locationText: { type: string, nullable: true }
                  description: { type: string, nullable: true }
                  mapUrl: { type: string, nullable: true }
                  meetupStatus: { type: string }
                  isRequested: { type: boolean }
              examples:
                sample:
                  summary: Example meetup details
                  value:
                    id: "2a7b0c5e-1a2b-4d3e-9f10-112233445566"
                    name: "Coffee chat"
                    startsAt: "2025-09-01T10:00:00.000Z"
                    endsAt: "2025-09-01T11:00:00.000Z"
                    city: "Kochi"
                    state: "Kerala"
                    country: "India"
                    activityId: 3
                    whoPays: "split"
                    feeAmount: "0"
                    guests: 2
                    hostId: "d7c351f8-18a4-4840-ac6f-2f27103cd048"
                    hostName: "Amogh"
                    hostAvatar: "https://cdn.example.com/avatars/amogh.jpg"
                    locationText: "Cafe XYZ, MG Road"
                    description: "Casual coffee to discuss tech and movies."
                    mapUrl: "https://maps.example.com/?q=cafe"
                    meetupStatus: "active"
                    isRequested: false
        '404':
          description: Not found
    patch:
      summary: Update a meetup
      security:
        - XUserId: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/UnauthorizedHeader' }
    delete:
      summary: Delete a meetup
      security:
        - XUserId: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /meetups/{id}/attendees:
    get:
      summary: List meetup attendees
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Attendees, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /meetups/{id}/requests:
    post:
      summary: Send request to join meetup
      security:
        - XUserId: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 500 }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/UnauthorizedHeader' }
  /meetups/me/requests/sent:
    get:
      summary: Sent meetup requests
      security:
        - XUserId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200': { description: Sent list, content: { application/json: { schema: { type: object } } } }
  /meetups/me/requests/received:
    get:
      summary: Received meetup requests
      security:
        - XUserId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200': { description: Received list, content: { application/json: { schema: { type: object } } } }
  /meetups/requests/{id}/approve:
    post:
      summary: Approve meetup request
      security:
        - XUserId: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Approved }
  /meetups/requests/{id}/decline:
    post:
      summary: Decline meetup request
      security:
        - XUserId: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Declined }
...